requirements.txt
grok
python-dotenv
streamlit